IMPALA-2945: Account for duplicate keys on multiple nodes preAgg
AggregationNode.computeStats() estimate cardinality under single node
assumption. This can be an underestimation in preaggregation node case
because same grouping key may exist in multiple nodes during
preaggreation.

This patch adjust the cardinality estimate using following model for the
number of distinct values in a random sample of k rows. Assumes we are
picking k rows from an infinite sample with ndv distinct values, with
the value uniformly distributed.

The probability of a given value not appearing in a sample, in that case
is

((NDV - 1) / NDV) ^ k

This is because we are making k choices, and each of them has
(ndv - 1) / ndv chance of not being our value. Therefore the
probability of a given value appearing in the sample is:

1 - ((NDV - 1) / NDV) ^ k

And the number of distinct values in the sample is:

(1 - ((NDV - 1) / NDV) ^ k) * NDV

This adjustment is done during distributed planning and assume the
minimum number of preaggregation instances per host. It intentionally
does not use getNumInstances() because it might change during parallel
planning phase or cost based planning phase. Therefore, preaggregation
node cardinality may still underestimate in case of MT_DOP > 1 and
COMPUTE_PROCESSING_COST = true.

Testing:
- Pass core tests.

Change-Id: I04c563e59421928875b340cb91654b9d4bc80b55