IMPALA-14101: Calcite planner: Add cost model calculations
This commit adds the cost model and calculations to be used in the join
optimizer rule. The ImpalaCost object implements the RelOptCost interface
and contains values which contribute to a cost.  The ImpalaCost object
roughly mirrors the Calcite VolcanoCost object with some slight variations.
The ImpalaCost object only looks at the cpu and io cost and ignores the
rowCount cost.

The ImpalaCost object is generated in the ImpalaRelMdNonCumulativeCost
class which is called from Calcite for a given RelNode. The cost generated
by this object uses the various inputs of the RelNode to calculate the
cpu and io time for the given logical node. Note that this is a
non-cumulative cost.  A cumulative cost exists within Calcite as well, but
there was no need to change the cumulative cost logic.

A TODO in this commit is to allow various cpu and io costs to be configurable.

Change-Id: I1e52b0e11e9a6d5814b0313117dd9c56602f3ff5