IMPALA-14102: Calcite Planner: Customize LoptOptimizeJoinRule
This commit provides a customized version of the LoptOptimizeJoinRule.

The ImpalaLoptOptimizeJoin class was grabbed from
https://github.com/apache/calcite/blob/calcite-1.37.0/core/src/main/java/org/apache/...
/calcite/rel/rules/LoptOptimizeJoinRule.java

The use of runtime filters within Impala throw a wrench into the join optimization
ordering. We should factor this in to come up with much better join ordering choices.
While this first attempt in this file does not do this, it does put some band-aids
on cases where the Calcite algorithm can go awry. Namely, the following changes
are made from the original Calcite algorithm:

- When attempting the check of whether to push down a join versus keeping the current
  join, the epsilon on the tiebreaker has been made a little wider. If it does go
  to the tiebreaker, it compares the "leftest most" cardinality on the 2 choices. This
  allows a more effective runtime filter to be used, if found.

- "Complex" trees through a bit of a wrench into runtime filtering. A "simple" tree is
  defined here as a tree with only projects and filters. If a tree is complex, the row
  count will have gone through some major transformation, and it isn't quite apparent
  based on the row count at the top level that a runtime filter could have a major
  impact. It especially caused problems when the complex tree was processed in between
  simple trees. Because the row count showed up low, a swap might occur with a simple
  tree where it shouldn't have. In practice, it helped extremely to delay the
  processing of complex trees until after all the simple trees have been processed.

- Similar to the last point, we also do not want to swap the left and right side
  if we detect a complex tree that has its "leftest" most input with a high
  cardinality because a runtime filter could help improve performance greatly.

To sum up: These are band-aid patches that take guesses as to when a runtime
filter will be used. A more complete analysis should be done at some point, but
these changes should improve performance, as tested on the tpcds queries.

Change-Id: I95bba7bafcc7dd86f17921c11740ac349daac6c7