IMPALA-14102: Calcite Planner: Pre-commit...
This commit is a pre-commit for IMPALA-14102. This jira customizes
the LoptOptimizeJoinRule.  This commit serves as a placeholder
starting point so we can easily see the customized changes that
are made by comparing the real IMPALA-14102 against the file in
this commit which is a direct copy of the rule file in Calcite v1.37.

Change-Id: I63daf6dacf0547a0488c1ecf0bc185b548e00d87